function/requirements.txt
azure-functions==1.2.1
psycopg2-binary==2.8.5
sendgrid==6.3.0
